requirements.txt
dj-database-url==0.5.0
Django==2.2.4
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.0
django-braces
django-embed-video
django-heroku
python-memcached
django-memcache-status
django_filters

